app/requirements.txt
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
django==5.2.7
djangorestframework==3.16.1
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
psycopg==3.2.11
psycopg-binary==3.2.11
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.3
urllib3==2.5.0
